最終課題その１

+++ practice/src/main/java/com/example/demo/repository/ContactRepository.java
@@ -4,6 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.entity.Contact;
 
-public interface ContactRepository extends JpaRepository<Contact, Long> {
- 
+public interface ContactRepository extends JpaRepository<Contact, Integer> {
 }
